requirements.txt
discord.py==2.3.2
python-dotenv
Flask
aiohttp
typing_extensions